Refactor folder structure in order to keep related functionality grouped
