+++ src/app/app-routing.module.ts
@@ -1,10 +1,10 @@
-import { AuthGuardService } from "./service/auth/auth-guard.service"
+import { AuthGuardService } from "./auth/auth-guard/auth-guard.service"
 import { AccountSettingsComponent } from "./account-settings/account-settings.component";
 import { DashboardComponent } from "./dashboard/dashboard.component";
-import { LoginComponent } from "./login/login.component";
-import { ProjectDetailComponent } from "./project-detail/project-detail.component";
-import { ProjectsListComponent } from "./projects-list/projects-list.component";
-import { UsersComponent } from "./users/users.component";
+import { LoginComponent } from "./auth/login/login.component";
+import { ProjectDetailComponent } from "./project/project-detail/project-detail.component";
+import { ProjectsListComponent } from "./project/projects-list/projects-list.component";
+import { UsersComponent } from "./admin/users/users.component";
 
 import { NgModule } from "@angular/core";
 import { Routes, RouterModule } from "@angular/router";
