+++ src/app/general-service/device-detector/device-detector.service.spec.ts
