+++ src/app/general-service/device-detector/device-detector.service.ts
