+++ src/app/project/project-detail/project-detail.component.scss
