+++ src/app/project/projects-list/projects-list.component.scss
