+++ src/app/project/projects-list/projects-list.component.ts
@@ -1,11 +1,11 @@
-import { Project } from "../model/project/project";
-import { University } from "../model/university/university";
-import { projectTypesList } from "../model/project/project-type";
-import { SortAttribute } from "../model/general/general.model";
+import { Project } from "../../model/project/project";
+import { University } from "../../model/university/university";
+import { projectTypesList } from "../../model/project/project-type";
+import { SortAttribute } from "../../model/general/general.model";
 
-import { ProjectsService } from "../service/projects/projects.service";
-import { UniversitiesService } from "../service/universities/universities.service";
-import { DeviceDetectorService } from "../service/device-detector/device-detector.service";
+import { ProjectsService } from "../projects-service/projects.service";
+import { UniversitiesService } from "../universities-service/universities.service";
+import { DeviceDetectorService } from "../../general-service/device-detector/device-detector.service";
 import { ProjectDetailComponent } from "../project-detail/project-detail.component";
 
 import { Router } from "@angular/router";
